app/API/requirements.txt
fastapi
uvicorn[standard]
pytest
httpx
openai
google-generativeai
pydantic-settings
pydantic
pytest-asyncio
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
aiosqlite>=0.17.0
aiofiles>=23.0.0
python-multipart>=0.0.6
qdrant-client>=1.7.0
tiktoken>=0.5.0